frontend/requirements.txt
streamlit
httpx
##########################################################################################
# EXAMPLE (Vision)
pillow
##########################################################################################
